Convert the percentage charge to a gfloat type.
2008-07-27  Richard Hughes  <richard@hughsie.com>

* src/gpm-cell-array.c: (gpm_cell_array_update),
(gpm_cell_percent_changed_cb), (gpm_cell_array_get_description),
(gpm_cell_array_class_init):
* src/gpm-cell-array.h:
* src/gpm-cell-unit.c: (gpm_cell_unit_init), (gpm_cell_unit_print),
(gpm_st_cell_unit):
* src/gpm-cell-unit.h:
* src/gpm-cell.c: (gpm_cell_refresh_hal_all),
(hal_device_property_modified_cb), (gpm_cell_set_phone_index),
(gpm_cell_get_description), (phone_device_refresh_cb),
(gpm_cell_class_init):
* src/gpm-cell.h:
* src/gpm-engine.c: (gpm_cell_array_percent_changed_cb),
(gpm_cell_array_charge_low_cb),
(gpm_cell_array_charge_critical_cb),
(gpm_cell_array_charge_action_cb):
* src/gpm-manager.c: (gpm_engine_charge_low_cb),
(gpm_engine_charge_critical_cb):
* src/gpm-tray-icon.c: (gpm_tray_icon_add_device):
Convert the percentage charge to a gfloat type.

svn path=/trunk/; revision=2851

+++ src/gpm-cell-array.h
@@ -47,7 +47,7 @@ typedef struct
 {
 	GObjectClass	parent_class;
 	void		(* percent_changed)	(GpmCellArray	*cell_array,
-						 guint		 percent);
+						 gfloat		 percent);
 	void		(* charging_changed)	(GpmCellArray	*cell_array,
 						 gboolean	 charging);
 	void		(* discharging_changed)	(GpmCellArray	*cell_array,
@@ -58,11 +58,11 @@ typedef struct
 	void		(* low_capacity)	(GpmCellArray	*cell_array,
 						 guint		 capacity);
 	void		(* charge_low)		(GpmCellArray	*cell_array,
-						 guint		 percent);
+						 gfloat		 percent);
 	void		(* charge_critical)	(GpmCellArray	*cell_array,
-						 guint		 percent);
+						 gfloat		 percent);
 	void		(* charge_action)	(GpmCellArray	*cell_array,
-						 guint		 percent);
+						 gfloat		 percent);
 	void		(* fully_charged)	(GpmCellArray	*cell_array);
 	void		(* collection_changed)	(GpmCellArray	*cell_array);
 	void		(* discharging)		(GpmCellArray	*cell_array);
